PI-33 Move widgets to a separate package
