+++ mobile/lib/screens/home_screen.dart
@@ -1,7 +1,7 @@
 import 'package:awesome_flutter_extensions/awesome_flutter_extensions.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:pi_mobile/app_navigation_drawer.dart';
+import 'package:pi_mobile/widgets/app_navigation_drawer.dart';
 import 'package:pi_mobile/provider/auth_provider.dart';
 
 class HomeScreen extends ConsumerWidget {
