+++ mobile/lib/widgets/navigation_drawer_entry.dart
